Restructure, add ExportHelper
