+++ src/client/java/com/skycatdev/autocut/AutocutCommandHandler.java
@@ -6,6 +6,9 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
+import com.skycatdev.autocut.export.ExportHelper;
+import com.skycatdev.autocut.record.OBSHandler;
+import com.skycatdev.autocut.record.RecordingManager;
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.text.Text;
